mytool/requirements.txt
typer
python-dotenv
rich
requests